# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zak Ecosystem AI - CRM Omnicanale API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calendar/google/sync:
    post:
      operationId: syncGoogleCalendar
      tags: [calendar]
      summary: Run Google Calendar synchronization
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleCalendarSyncInput"
      responses:
        "200":
          description: Synchronization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleCalendarSyncResult"
...
  /webhook/whatsapp:
    get:
      operationId: webhookWhatsappChallenge
      tags: [webhooks]
      summary: Verify Meta WhatsApp webhook challenge
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge accepted
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Invalid verify token
    post:
      operationId: webhookWhatsapp
      tags: [webhooks]
      summary: Receive WhatsApp message from Meta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
  /webhook/voice-assistant:
    post:
      operationId: webhookVoiceAssistant
      tags: [webhooks]
      summary: Receive generic, Vapi, or Bland voice call transcription webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceWebhookPayload"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
  /webhook/google-calendar:
    post:
      operationId: webhookGoogleCalendar
      tags: [webhooks]
      summary: Receive Google Calendar push notification and trigger incremental sync
      responses:
        "200":
          description: Notification accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Invalid Google channel token
components:
  schemas:
...
    GoogleCalendarSyncInput:
      type: object
      properties:
        direction:
          type: string
          enum: [zak_to_google, google_to_zak, bidirectional]
          default: bidirectional
        full_sync:
          type: boolean
          default: false
        days_ahead:
          type: integer
          default: 180
    GoogleCalendarSyncResult:
      type: object
      required: [configured, direction, pushed, pulled, skipped, conflicts, errors]
      properties:
        configured:
          type: boolean
        direction:
          type: string
        pushed:
          type: integer
        pulled:
          type: integer
        skipped:
          type: integer
        conflicts:
          type: integer
        errors:
          type: array
          items:
            type: string
        message:
          type: string
...
    WebhookPayload:
      type: object
      required: [object]
      properties:
        object:
          type: string
        entry:
          type: array
          items:
            type: object
    VoiceWebhookPayload:
      type: object
      properties:
        provider:
          type: string
          enum: [generic, vapi, bland]
        trascrizione:
          type: string
        telefono:
          type: ["string", "null"]
          description: If present, the webhook links the call to an existing CRM contact with the same normalized phone
        durata:
          type: ["integer", "null"]
        call_id:
          type: ["string", "null"]
        recording_url:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        message:
          type: object
          description: Raw Vapi webhook message payload.
        call:
          type: object
          description: Raw Bland or generic call payload.
        data:
          type: object
          description: Raw provider-specific payload wrapper.
        analysis:
          type: object
          description: Provider intent analysis or extracted variables.

    # ── AUTOMAZIONI ────────────────────────────────────────────────────────────
...
    WebhookResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]